refactor: 컨벤션에 맞게 모임을 칭하는 용어를 story -> social 으로 변경
